move classes into individual files in biosnicar/classes
