+++ biosnicar/classes/__init__.py
@@ -1,0 +1,19 @@
+from .impurity import Impurity
+from .ice import Ice
+from .illumination import Illumination
+from .rt_config import RTConfig
+from .model_config import ModelConfig
+from .outputs import Outputs
+from .plot_config import PlotConfig
+from .bio_optical_config import BioOpticalConfig
+
+__all__ = [
+    'Impurity',
+    'Ice',
+    'Illumination',
+    'RTConfig',
+    'ModelConfig',
+    'Outputs',
+    'PlotConfig',
+    'BioOpticalConfig'
+]
